fix: 🐞 fixing bugs due to components refactoring
