+++ packages/frontend/src/components/pages/GeneralHome/GeneralHome.view.jsx
@@ -1,346 +1,17 @@
-import { useState } from 'react';
-
-import { styled } from '@mui/material/styles';
-
-import { makeStyles } from '@material-ui/core';
+import React from 'react';
 import Chip from '@mui/material/Chip';
 import Stack from '@mui/material/Stack';
 import TabsUnstyled from '@mui/base/TabsUnstyled';
-import TabsListUnstyled from '@mui/base/TabsListUnstyled';
-import TabPanelUnstyled from '@mui/base/TabPanelUnstyled';
-import buttonUnstyledClasses from '@mui/base/ButtonUnstyled';
-import TabUnstyled, { tabUnstyledClasses } from '@mui/base/TabUnstyled';
 import Grid from '@mui/material/Grid';
-
 import Helmet from 'react-helmet';
 import DayPicker from 'react-day-picker';
 import 'react-day-picker/lib/style.css';
-
 import Carousel from 'react-material-ui-carousel';
 
-const useStyles = makeStyles({
-	header: {
-		fontWeight: 700,
-	},
-	mainWrapper: {
-		display: 'flex',
-		flexDirection: 'column',
-		justifyContent: 'center',
-		alignItems: 'center',
-		backgroundColor: '#F8F8F8',
-	},
-	horizontalScrollWrapper: {
-		'display': 'flex',
-		'minHeight': '7.9vh',
-		'overflow': 'hidden',
-		'&:hover': {
-			'overflowX': 'scroll',
-			'&::-webkit-scrollbar-track': {
-				borderRadius: '10px',
-				backgroundColor: '#F3F4F8',
-			},
-			'&::-webkit-scrollbar': {
-				borderRadius: '10px',
-				height: '7px',
-				zIndex: 2,
-				backgroundColor: '#F5F5F5',
-			},
-			'&::-webkit-scrollbar-thumb': {
-				borderRadius: '10px',
-				backgroundColor: '#D5DBE7',
-			},
-		},
-	},
-	mostClickedIdeasWrapper: {
-		display: 'flex',
-		justifyContent: 'space-between',
-		alignItems: 'center',
-		backgroundColor: '#EDF2FB',
-		padding: '0 10px 0 10px',
-		margin: '10px',
-		borderRadius: '8px',
-		minWidth: '25.2vh',
-		maxHeight: '6vh',
-		animation: 'mostClickedIdeas 10s -10s linear infinite',
-		willChange: 'transform',
-	},
-	mostClickedIdeasTitle: {
-		color: '#8197ae',
-		fontSize: '.9rem',
-	},
-	mostClickedIdeasContent: {
-		color: '#151515',
-		fontSize: '1rem',
-	},
-	carousel: {
-		borderRadius: '8px',
-		backgroundColor: '#fff',
-		border: '2px solid #EDEEF1',
-		padding: '8px',
-	},
-	carouselContect: {
-		maxHeight: '6vh',
-		textAlign: 'center',
-		fontSize: '1.1rem',
-		padding: '0 3% 0 3%',
-	},
-	focusIdeas: {
-		padding: '10px',
-		borderRadius: '8px',
-		backgroundColor: '#fff',
-		border: '2px solid #EDEEF1',
-	},
-	focusIdeasWrapper: {
-		display: 'flex',
-		flexDirection: 'column',
-		padding: '12px',
-		animation: 'latestNews 20s -20s linear infinite',
-		marginTop: '10px',
-		borderRadius: '8px',
-		backgroundColor: '#F3F4F8',
-	},
-	focusIdeasScroll: {
-		'display': 'flex',
-		'flexDirection': 'column',
-		'minHeight': '36vh',
-		'maxHeight': '36vh',
-		'overflow': 'hidden',
-		'&:hover': {
-			'overflowY': 'scroll',
-			'&::-webkit-scrollbar-track': {
-				borderRadius: '10px',
-				backgroundColor: '#F3F4F8',
-			},
-			'&::-webkit-scrollbar': {
-				borderRadius: '10px',
-				maxWidth: '7px',
-				zIndex: 2,
-				backgroundColor: '#F5F5F5',
-			},
-			'&::-webkit-scrollbar-thumb': {
-				borderRadius: '10px',
-				backgroundColor: '#D5DBE7',
-			},
-		},
-	},
-	focusIdeasDate: {
-		fontSize: '.9rem',
-		color: '#8197ae',
-	},
-	focusIdeasContent: {
-		color: '#8197ae',
-		fontSize: '1.04rem',
-		marginTop: '10px',
-	},
-	lastPublications: {
-		borderRadius: '8px',
-		backgroundColor: '#fff',
-		border: '2px solid #EDEEF1',
-		padding: '10px',
-		marginBottom: '25px',
-	},
-	lastPublicationsWrapper: {
-		display: 'flex',
-		flexDirection: 'column',
-		backgroundColor: '#EDF2FB',
-		padding: '12px',
-		borderRadius: '8px',
-		marginTop: '10px',
-	},
-	lastPublicationsTitle: {
-		color: '#8197ae',
-		fontSize: '.9rem',
-	},
-	lastPublicationsContent: {
-		color: '#151515',
-		fontWeight: 'bold',
-		fontSize: '1.1rem',
-		marginTop: '10px',
-	},
-	industryRecoursed: {
-		marginBottom: '25px',
-	},
-	industryRecoursedWrapper: {
-		display: 'flex',
-		flexDirection: 'column',
-		padding: '12px',
-		marginTop: '10px',
-		borderRadius: '8px',
-		border: '2px solid #EDEEF1',
-		backgroundColor: '#fff',
-	},
-	industryRecoursedDate: {
-		fontSize: '.9rem',
-		color: '#8197ae',
-	},
-	industryRecoursedContent: {
-		color: '#151515',
-		fontSize: '1.04rem',
-		marginTop: '10px',
-	},
-	latestNews: {
-		padding: '10px',
-		marginBottom: '25px',
-		borderRadius: '8px',
-		backgroundColor: '#fff',
-		border: '2px solid #EDEEF1',
-	},
-	latestNewsScroll: {
-		'display': 'flex',
-		'flexDirection': 'column',
-		'minHeight': '20.1vh',
-		'maxHeight': '20.1vh',
-		'overflow': 'hidden',
-		'&:hover': {
-			'overflowY': 'scroll',
-			'&::-webkit-scrollbar-track': {
-				borderRadius: '10px',
-				backgroundColor: '#F3F4F8',
-			},
-			'&::-webkit-scrollbar': {
-				borderRadius: '10px',
-				maxWidth: '7px',
-				zIndex: 2,
-				backgroundColor: '#F5F5F5',
-			},
-			'&::-webkit-scrollbar-thumb': {
-				borderRadius: '10px',
-				backgroundColor: '#D5DBE7',
-			},
-		},
-	},
-	latestNewsWrapper: {
-		'display': 'flex',
-		'flexDirection': 'column',
-		'paddingBottom': '4px',
-		'margin': '5px',
-		'minWidth': '25.2vh',
-		'animation': 'latestNews 25s -25s linear infinite',
-		'willChange': 'transform',
-		'&:not(:last-child)': {
-			borderBottom: '1px solid #EDEDF0',
-		},
-	},
-	latestNewsHeader: {
-		fontSize: '.9rem',
-		color: '#8197ae',
-	},
-	latestNewsContent: {
-		color: '#151515',
-		fontSize: '1.04rem',
-		marginTop: '10px',
-	},
-	morningNotes: {
-		borderRadius: '8px',
-		backgroundColor: '#fff',
-		border: '2px solid #EDEEF1',
-		padding: '10px',
-	},
-	morningNotesWrapper: {
-		'display': 'flex',
-		'flexDirection': 'column',
-		'paddingBottom': '4px',
-		'margin': '5px',
-		'minWidth': '25.2vh',
-		'animation': 'latestNews 30s -30s linear infinite',
-		'willChange': 'transform',
-		'&:not(:last-child)': {
-			borderBottom: '1px solid #EDEDF0',
-		},
-	},
-	morningNotesScroll: {
-		'display': 'flex',
-		'flexDirection': 'column',
-		'minHeight': '19.5vh',
-		'maxHeight': '19.5vh',
-		'marginTop': '2px',
-		'overflow': 'hidden',
-		'&:hover': {
-			'overflowY': 'scroll',
-			'&::-webkit-scrollbar-track': {
-				borderRadius: '10px',
-				backgroundColor: '#F3F4F8',
-			},
-			'&::-webkit-scrollbar': {
-				borderRadius: '10px',
-				maxWidth: '7px',
-				zIndex: 2,
-				backgroundColor: '#F5F5F5',
-			},
-			'&::-webkit-scrollbar-thumb': {
-				borderRadius: '10px',
-				backgroundColor: '#D5DBE7',
-			},
-		},
-	},
-	morningNotesDate: {
-		fontSize: '.9rem',
-		color: '#8197ae',
-	},
-	morningNotesContent: {
-		fontSize: '.9rem',
-		color: '#151515',
-	},
-	events: {
-		borderRadius: '8px',
-		backgroundColor: '#fff',
-		border: '2px solid #EDEEF1',
-		padding: '10px',
-	},
-	eventsWrapper: {
-		display: 'flex',
-		flexDirection: 'column',
-		padding: '12px',
-		marginTop: '10px',
-		borderTop: '2px solid #EDEDF0',
-	},
-	eventsInnerWrapper: {
-		'display': 'flex',
-		'justifyContent': 'center',
-		'&:not(:last-child)': {
-			marginBottom: '15px',
-		},
-	},
-	eventsLabel: {
-		padding: '5px',
-		marginRight: '10px',
-		borderRadius: '0 6px 6px 0',
-	},
-	eventsContentWrapper: {
-		display: 'flex',
-		flexDirection: 'column',
-		width: '100%',
-	},
-	eventsInnerContentWrapper: {
-		'display': 'flex',
-		'flexDirection': 'row',
-		'justifyContent': 'space-between',
-		'&:not(:last-child)': {
-			marginBottom: '6px',
-		},
-	},
-	eventsHeader: {
-		color: '#0F0F0F',
-		fontWeight: 'bold',
-		fontSize: '.9rem',
-	},
-});
+import useStyles, {Tab, TabPanel,TabsList} from './GeneralHome.style';
 
-const home = () => {
+const GeneralHomeView = (props) => {
 	const classes = useStyles();
-
-	const date = new Date();
-
-	const calendarDefaultValue = {
-		year: date.getFullYear(),
-		month: date.getMonth(),
-		day: date.getDay(),
-	};
-
-	const [selectedDay, setSelectedDay] = useState(calendarDefaultValue);
-
-	const events2 = [1, 2, 3, 4, 5, 6, 7];
-
 	const renderDay = (day) => {
 		const dates = day.getDate();
 		const dateStyle = {
@@ -354,7 +25,7 @@ const home = () => {
 		return (
 			<div style={cellStyle}>
 				<div style={dateStyle}>{dates}</div>
-				{events2.includes(dates) ? (
+				{props.events2.includes(dates) ? (
 					<div
 						style={{
 							display: 'flex',
@@ -1437,8 +1108,8 @@ const home = () => {
 									</Helmet>
 									<DayPicker
 										renderDay={renderDay}
-										selectedDays={selectedDay}
-										onDayClick={setSelectedDay}
+										selectedDays={props.selectedDay}
+										onDayClick={props.setSelectedDay}
 									/>
 								</div>
 								<section className={classes.eventsWrapper}>
@@ -1556,52 +1227,7 @@ const home = () => {
 	);
 };
 
-export default home;
+GeneralHomeView.displayName = 'MemberHomeView';
+GeneralHomeView.defaultProps = {};
 
-const Tab = styled(TabUnstyled)`
-	font-family: IBM Plex Sans, sans-serif;
-	color: #868da2;
-	cursor: pointer;
-	font-size: 0.875rem;
-	background-color: transparent;
-	width: 100%;
-	border: none;
-	border-radius: 30px;
-	display: flex;
-	justify-content: center;
-
-	&:hover {
-		background-color: #1c67ff;
-		padding: 6px 6px;
-	}
-
-	&:focus {
-		border-radius: 30px;
-		outline-offset: 2px;
-	}
-
-	&.${tabUnstyledClasses.selected} {
-		background-color: #1c67ff;
-		padding: 6px 6px;
-		color: #fff;
-	}
-
-	&.${buttonUnstyledClasses.disabled} {
-		opacity: 0.5;
-		cursor: not-allowed;
-	}
-`;
-
-const TabPanel = styled(TabPanelUnstyled)`
-	width: 100%;
-	font-size: 0.875rem;
-`;
-
-const TabsList = styled(TabsListUnstyled)`
-	background-color: #e2ebfc;
-	border-radius: 30px;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	align-content: space-between;
-`;
+export default React.memo(GeneralHomeView);
